pyproject.toml
[project]
name = "airflow-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow==2.9.0",
    "bs4>=0.0.2",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.11",
]
